images: adding a history of all past splashes.
Recreated through good ol' git archeology. It is hopefully more or less
complete. It even contains splashes which made it to no releases.

This will be the new procedure when creating a new splash: just adding a link to
easily find it back and displaying it as an image in the file for quick
scrolling of splashes.