image: adding info and links on wilber-tips and wilber-wizard images.
These are not used in GIMP so there is no need to keep them in the main repo,
nor is it needed to add them to the gimp-data repo. Nevertheless for historical
(and fun) reason, we want to keep a trace of them (not just lost in `git log`
where it won't likely ever reappear). Showcasing them in our log of historical
Wilber is the right way to do it.

Also adding some dates information as it's interesting.