meson: install icons with meson.
This commit is separate from the previous to make it immediately clear (by
comparing files) that the previous commit is a simple move with no modification
whatsoever of the icons/ directory, i.e. the symmetrical add of removal in MR
gimp!1373.

This also fixes the dialog-question and gimp-wilber-eek icons being generated as
256×256 PNG images instead of 64×64 in the vector icon cases (where only these 2
files are generated as PNG, additionally to all the SVG icons) and which went
unnoticed until we moved the meson to gimp-data.