images: add a README specifically for this folder.
Also add a missing image display (the link was present, but the image was not
displayed) in the logo history.

For anyone wondering why I'm not already pushing the images themselves: it's
because we are expecting new ones soon, so let's start with the new ones instead
of pushing soon obsolete ones.