images: generate gimp.ico on demand.
This rule is meant to replace build/windows/wilber.ico currently committed in
the main repository. Note that it is not made to be run automatically because it
has a circular dependency with GIMP itself:

* GIMP requires this .ico file (it is referenced in `build/windows/gimp.rc`
  which is compiled in the main binary through windows.compile_resources() in
  `app/meson.build`) at compilation time.
* We generate the .ico file through GIMP scripting.

As a consequence, we will still commit this file as image/gimp.ico, yet adding
an update procedure in images/README.md which needs to be run when updating the
logo (or if any rule changes for .ico creation, e.g. regarding sizes recommended
by Microsoft).